# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  domain.PaymentOrderStatusRequest:
    properties:
      booking_id:
        type: integer
      order_id:
        type: string
    required:
    - booking_id
    - order_id
    type: object
...
  domain.PaymentStatusRequest:
    properties:
      payment_id:
        type: string
    required:
    - payment_id
    type: object
...
  domain.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  http.CreateSettingRequest:
    properties:
      data_type:
        enum:
        - string
        - integer
        - decimal
        - boolean
        type: string
      description:
        type: string
      is_active:
        type: boolean
      setting_key:
        type: string
      setting_value:
        type: string
    required:
    - data_type
    - setting_key
    - setting_value
    type: object
...
  http.UpdateSettingRequest:
    properties:
      data_type:
        enum:
        - string
        - integer
        - decimal
        - boolean
        type: string
      description:
        type: string
      is_active:
        type: boolean
      setting_value:
        type: string
    required:
    - data_type
    - setting_value
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для платформы аренды квартир Renti.kz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Renti.kz API
  version: "1.0"
paths:
...
  /payments/check-status:
    post:
      consumes:
      - application/json
      description: Проверяет статус платежа в системе FreedomPay по payment_id
      parameters:
      - description: Payment ID для проверки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.PaymentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Проверка статуса платежа
      tags:
      - payments
  /payments/check-status-by-order:
    post:
      consumes:
      - application/json
      description: Проверяет статус платежа в системе FreedomPay по order_id
      parameters:
      - description: Order ID и Booking ID для проверки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.PaymentOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Проверка статуса платежа по Order ID
      tags:
      - payments
  /settings:
    get:
      consumes:
      - application/json
      description: Возвращает все настройки платформы (только для администраторов)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение всех настроек
      tags:
      - settings
    post:
      consumes:
      - application/json
      description: Создает новую настройку платформы (только для администраторов)
      parameters:
      - description: Данные настройки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateSettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создание новой настройки
      tags:
      - settings
  /settings/{key}:
    delete:
      consumes:
      - application/json
      description: Удаляет настройку платформы (только для администраторов)
      parameters:
      - description: Ключ настройки
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление настройки
      tags:
      - settings
    get:
      consumes:
      - application/json
      description: Возвращает конкретную настройку по ключу (только для администраторов)
      parameters:
      - description: Ключ настройки
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение настройки по ключу
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Обновляет существующую настройку платформы (только для администраторов)
      parameters:
      - description: Ключ настройки
        in: path
        name: key
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновление настройки
      tags:
      - settings
  /settings/service-fee:
    get:
      consumes:
      - application/json
      description: Возвращает текущий процент сервисного сбора платформы
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Получение процента сервисного сбора
      tags:
      - settings
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"